Many small fixes to make code more ANSI-C and clean up warnings. Mainly changes from char * to const char *.

+++ core/alsp_src/generic/alloc.c
@@ -195,7 +195,7 @@ double_val(pword p)
 	int i;
 	
 	for (i = 0; i < 4; i++) {
-	  	*(((short *) &dblval) + i) = (short) TERM_ARGN(p, i + 1);
+	  	*(((short *) &dblval) + i) = (long) TERM_ARGN(p, i + 1);
 	}
 	return dblval;
 }
